Release v0.3.0 from Develop
Move into the Solid Fuel infrastructure

Add the installers (windows, macos, and debian)

fix bugs

CHange name to solidArp

commit b6fa0161e3146470d3576e284c01dac6432e4c50
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Wed Dec 6 14:50:46 2023 -0500

    update docs

commit 7de03346ca3d3ff8e4b8dc2e96c94a538417bc0c
Merge: d068b63 1670b27
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Wed Dec 6 14:48:24 2023 -0500

    Merge remote-tracking branch 'refs/remotes/origin/Develop' into Develop

commit d068b6333d69cac136bdc6c586c39144c89bdb1c
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Wed Dec 6 14:47:42 2023 -0500

    update commons

commit 1670b274e4b947b2494c155ddf6097755e059f2c
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Tue Dec 5 17:44:28 2023 -0500

    clean ups

commit f76c35d0c74dfb18a7a5a7e2efd6f98c268f1ea6
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Tue Dec 5 15:48:02 2023 -0500

    relax the concurrency groups

commit a9b1296fc0e60cd859c57938888611fc5583aab3
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Tue Dec 5 15:42:50 2023 -0500

    add action for pluginval

commit b7cea913052ebea394ca9b2e8411714023983c71
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Tue Dec 5 14:59:06 2023 -0500

    put back pull_request event

commit 3285f72336e2918a3f2139d36f495a9cfa496a7b
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Tue Dec 5 14:55:02 2023 -0500

    give required ref in uses clause

commit e3277b1560dfd48af221fe6a1bdb5eceed222213
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Tue Dec 5 14:52:05 2023 -0500

    refactor the workflows a little

commit b9d33b026d7f32035684816a97573e90b1e1ffc0
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Mon Dec 4 20:08:12 2023 -0500

    Fix issue #22

commit 26edcf15f60f1978e6ee984c8f133deca80c4f2b
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Mon Dec 4 13:54:46 2023 -0500

    Installation guide

    But maybe this should go in common ?

commit 809d273506e26d7271a3ba63f95832d7b4abfc33
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Mon Dec 4 13:46:07 2023 -0500

    add workflow for manuals

commit bbe314b89821ea7591d37ec31d7b697a72c0b4ae
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Mon Dec 4 10:38:04 2023 -0500

    updates to build clean on MacOS

commit 54cc388683268e4e12b3083eed1e6d747db0a533
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sun Dec 3 13:38:13 2023 -0500

    integrate mac workflow

commit 2032d12df0511c31cbb9e1abd0c24844818f3608
Merge: 50ee5f2 fc0a6f4
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sun Dec 3 12:45:12 2023 -0500

    Merge branch 'linux-build' into Develop

commit 50ee5f24fa64bb60d10341245b2f2075dae11d91
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sun Dec 3 12:44:57 2023 -0500

    common

commit fc0a6f407e37b3da81415c158ea2ef883b527edf
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sun Dec 3 10:55:38 2023 -0500

    add the change log

commit 99ba00c9937b5593b1f552523c93345cf9776460
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sun Dec 3 10:32:26 2023 -0500

    fixups for the deb file

commit abec2354d3ace4f0c1cff60e42ca0e5c9258c73b
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sun Dec 3 10:18:04 2023 -0500

    forgot some

commit 462414c221b80e32cdfb17423b8e97e2134e2753
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sun Dec 3 10:14:36 2023 -0500

    integrate with workflows

commit 6706f22e7f056c78155e746625ae495dfd2fbe26
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sun Dec 3 09:44:12 2023 -0500

    integrate with workflow

commit 41777d0774f8ee4ab7329c3ec9bf12887b783787
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sat Dec 2 19:54:02 2023 -0500

    remember to checkout submodules

commit 01b160daca28cbd0b602e5addf54446f030da6aa
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sat Dec 2 19:51:36 2023 -0500

    fix up

commit 8777cd4113a04794131b438ba3392ba0f5b3bd97
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sat Dec 2 19:49:53 2023 -0500

    initegrate win build tools

commit 3fe0dc011201a934af160126d2cb363d2a2af452
Author: Mark Hollomon <mhhollomon@gmail.com>
Date:   Sat Dec 2 16:54:18 2023 -0500

    Initial intake to solidfuel

https://tonaldrift.com/releases/flotsam/#

+++ external/TinySHA1/TinySHA1.hpp
@@ -140,9 +140,9 @@ namespace sha1
 		void processBlock() {
 			uint32_t w[80];
 			for (size_t i = 0; i < 16; i++) {
-				w[i]  = (m_block[i*4 + 0] << 24);
-				w[i] |= (m_block[i*4 + 1] << 16);
-				w[i] |= (m_block[i*4 + 2] << 8);
+				w[i]  = (uint32_t(m_block[i*4 + 0]) << 24);
+				w[i] |= (uint32_t(m_block[i*4 + 1]) << 16);
+				w[i] |= (uint32_t(m_block[i*4 + 2]) << 8);
 				w[i] |= (m_block[i*4 + 3]);
 			}
 			for (size_t i = 16; i < 80; i++) {
